requirements.txt
openai>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.1
pyotp>=2.9.0
python-telegram-bot>=20.0
